Rename files and shuffle the order
